webapp/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
protobuf==3.19.1
six==1.16.0
Werkzeug==2.0.2
PyMySQL==1.0.2
requests==2.26.0